+++ CHANGELOG.md
@@ -44,6 +44,7 @@
 * Lua file automated reload supports editors that replace the file
 * Try supporting swr_alloc_set_opts2 even when compiled with old library
 * Enforce mesa software rendering even when nvidia driver is loaded
+* Catch when ffmpeg process exited
 
 ### Changed